refactor: 챌린지 상세 페이지 폴더 정리
