+++ src/app/home/challenge/[challengeId]/_componenets/ChallengeInfo/ChallengeInfo.type.ts
