+++ src/app/home/challenge/[challengeId]/_componenets/ChallengeStateNotice/ChallengeStateNotice.type.ts
