+++ src/app/home/challenge/[challengeId]/page.tsx
@@ -1,5 +1,5 @@
-import Header from "./_componenets/Header";
-import Main from "./_componenets/Main";
+import Header from "./_componenets/Header/Header";
+import Main from "./_componenets/Main/Main";
 
 export default function Challenge() {
     return (
